Add smarter logic for handling team joins
HandleTeamJoins used to simply wait 55 seconds before sending a
welcome message to a new user. We just hoped that the user would
have joined a team by then. Once we enabled requiring email
confirmation for new Mattermost users, this broke because most
users weren't getting their email confirmation message within those
55 seconds.

This commit eschews the "wait 55 seconds" method for a loop that
checks every 5 seconds if the user has joined a team, and if it has,
sends the welcome message, etc.

Note: An important distinction here is between a new user being
created—which happens after inputting their information on the
sign-up page—and a new user joining a team—which happens after email
confirmation and selecting the team to join.