Escape entire message when sending source
Before, when requesting the source of a message with three backticks
deliminating code, the code section of the message would not be
plaintext but would be instead formatted by Mattermost according to
Markdown formatting. This was because the plaintext section of holobot's
response was made using three backticks, so additional backticks
canceled those out.

This commit changes the method of creating the plaintext section of
holobot's message to using four spaces at the beginning of each line.
This method escapes all characters including backticks, and even
preserves other indenting and code.

Now, if you request a source of a message, copy the source, and then
paste it and post it again, your posted message should be identical to
the original message.

(Fixes #10)